+++ gmm_net/models/gmm_net_for_own_opp_team_performance.py
@@ -1,5 +1,5 @@
 import numpy as np
-from gmm_net.models.own_opp_team_regressor import OwnTeamOppTeamRegressor
+from gmm_net.models.base_gmm_net import OwnTeamOppTeamRegressor
 from gmm_net.models.unsupervised_kernel_regression import UnsupervisedKernelRegression as UKR
 from gmm_net.models.ukr_for_kde import UKRForWeightedKDE as UKRKDE
 from gmm_net.models.gplvm import GPLVM
